move initial context to session, not config
