+++ codex-rs/core/src/codex_delegate.rs
@@ -41,6 +41,7 @@ use crate::config::Config;
 use crate::guardian::GuardianApprovalRequest;
 use crate::guardian::review_approval_request_with_cancel;
 use crate::guardian::routes_approval_to_guardian;
+use crate::initial_context::InitialContextInclusions;
 use crate::mcp_tool_call::MCP_TOOL_APPROVAL_ACCEPT;
 use crate::mcp_tool_call::MCP_TOOL_APPROVAL_ACCEPT_FOR_SESSION;
 use crate::mcp_tool_call::MCP_TOOL_APPROVAL_DECLINE_SYNTHETIC;
@@ -63,6 +64,7 @@ use crate::codex::completed_session_loop_termination;
 #[allow(clippy::too_many_arguments)]
 pub(crate) async fn run_codex_thread_interactive(
     config: Config,
+    initial_context_inclusions: InitialContextInclusions,
     auth_manager: Arc<AuthManager>,
     models_manager: Arc<ModelsManager>,
     parent_session: Arc<Session>,
@@ -76,6 +78,7 @@ pub(crate) async fn run_codex_thread_interactive(
 
     let CodexSpawnOk { codex, .. } = Codex::spawn(CodexSpawnArgs {
         config,
+        initial_context_inclusions,
         auth_manager,
         models_manager,
         environment_manager: Arc::new(EnvironmentManager::from_environment(
@@ -171,6 +174,7 @@ pub(crate) async fn run_codex_thread_one_shot(
     let child_cancel = cancel_token.child_token();
     let io = run_codex_thread_interactive(
         config,
+        InitialContextInclusions::full(),
         auth_manager,
         models_manager,
         parent_session,
