+++ codex-rs/core/src/config/config_tests.rs
@@ -4562,7 +4562,6 @@ fn test_precedence_fixture_with_o3_profile() -> std::io::Result<()> {
             include_permissions_instructions: true,
             include_apps_instructions: true,
             include_environment_context: true,
-            initial_context_inclusions: InitialContextInclusions::full(),
             compact_prompt: None,
             commit_attribution: None,
             forced_chatgpt_workspace_id: None,
@@ -4710,7 +4709,6 @@ fn test_precedence_fixture_with_gpt3_profile() -> std::io::Result<()> {
         include_permissions_instructions: true,
         include_apps_instructions: true,
         include_environment_context: true,
-        initial_context_inclusions: InitialContextInclusions::full(),
         compact_prompt: None,
         commit_attribution: None,
         forced_chatgpt_workspace_id: None,
@@ -4856,7 +4854,6 @@ fn test_precedence_fixture_with_zdr_profile() -> std::io::Result<()> {
         include_permissions_instructions: true,
         include_apps_instructions: true,
         include_environment_context: true,
-        initial_context_inclusions: InitialContextInclusions::full(),
         compact_prompt: None,
         commit_attribution: None,
         forced_chatgpt_workspace_id: None,
@@ -4988,7 +4985,6 @@ fn test_precedence_fixture_with_gpt5_profile() -> std::io::Result<()> {
         include_permissions_instructions: true,
         include_apps_instructions: true,
         include_environment_context: true,
-        initial_context_inclusions: InitialContextInclusions::full(),
         compact_prompt: None,
         commit_attribution: None,
         forced_chatgpt_workspace_id: None,
