+++ codex-rs/core/src/lib.rs
@@ -38,6 +38,7 @@ mod flags;
 mod git_info_tests;
 mod guardian;
 mod hook_runtime;
+mod initial_context;
 mod installation_id;
 pub(crate) mod instructions;
 pub(crate) mod landlock;
